Split out functional tests
Update tests to use Table Driven Tests.
